ENH: Rename itol to plot.py
